Add AssumeRole support
Based on the sample AssumeRole code here:
https://docs.aws.amazon.com/sdk-for-go/api/aws/credentials/stscreds/

I've added an additional optional flag for a Role ARN.
If specified, I used the configuration (with our without a profile)
to request temporary credentials via the STS service.

These credentials are then passed to the DynamoDB and Kinesis modules.

Not really sure why a single session isn't created at the start
and passed to the other modules, but I'll leave that optimization
for another developer to fix